refactor code move download functions
